requirements.txt
attrs==21.2.0
autopep8==1.6.0
click==8.0.3
coverage==6.1.1
cycler==0.11.0
Flask==2.0.2
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
numpy==1.21.3
packaging==21.2
Pillow==8.4.0
pluggy==1.0.0
py==1.10.0
pycodestyle==2.8.0
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
scipy==1.7.1
six==1.16.0
toml==0.10.2
Werkzeug==2.0.2
